+++ packages/batch-submitter/src/batch-submitter/tx-batch-submitter.ts
@@ -52,13 +52,8 @@ export class TransactionBatchSubmitter extends BatchSubmitter {
       process.exit(1)
     }
     this.syncing = info.syncing
-
-    const addressManager = (
-      await getContractFactory('Lib_AddressManager', this.signer)
-    ).attach(info.addresses.addressResolver)
-    const ctcAddress = await addressManager.getAddress(
-      'OVM_CanonicalTransactionChain'
-    )
+    const addrs = await this._getChainAddresses(info)
+    const ctcAddress = addrs.ctcAddress
 
     if (
       typeof this.chainContract !== 'undefined' &&
